@@ -1,11 +1,11 @@
 name: sber_pay
 description: Flutter плагин для нативного платежного сервиса SberPay.
-version: 1.2.0
+version: 3.0.0
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
-  flutter: ^3.10.0
+  sdk: ">=3.4.0 <4.0.0"
+  flutter: ">=3.22.0"
 
 flutter:
   assets:
@@ -27,17 +27,10 @@ dependencies:
   flutter:
     sdk: flutter
   sber_pay_android:
-    git:
-      url: https://github.com/mkhamat/spay_sdk_flutter.git
-      ref: feat/v2.2.0
-      path: sber_pay_android
+    path: ../sber_pay_android
   sber_pay_ios:
-      git:
-        url: https://github.com/mkhamat/spay_sdk_flutter.git
-        ref: feat/v2.2.0
-        path: sber_pay_ios
+    path: ../sber_pay_ios
   sber_pay_platform_interface:
-      git:
-        url: https://github.com/mkhamat/spay_sdk_flutter.git
-        ref: feat/v2.2.0
-        path: sber_pay_platform_interface
+    path: ../sber_pay_platform_interface
+dev_dependencies:
+  flutter_lints: ^6.0.0
