@@ -1,12 +1,14 @@
 name: sber_pay_platform_interface
 description: A common platform interface for the sber_pay plugin.
-version: 1.2.0
+version: 3.0.0
 publish_to: none
 
 environment:
-  sdk: ">=3.0.0 <4.0.0"
+  sdk: ">=3.4.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  plugin_platform_interface: ^2.1.6
+  plugin_platform_interface: ^2.1.8
+dev_dependencies:
+  flutter_lints: ^6.0.0
