@@ -10,18 +10,21 @@ class SberPayPlugin {
   /// Инициализация SberPay SDK.
   ///
   /// Необходимо выполнить для начала работы с библиотекой.
-  /// На платформе Android этот метод является асинхронным, однако у
-  /// него нет API (коллбека) для выполнения кода после завершения
-  /// инициализации.
   ///
+  /// * [apiKey] - ключ, выдаваемый по договору
+  /// * [merchantLogin] - логин, выдаваемый по договору
   /// * [env] - среда запуска, которая определяется через [SberPayEnv].
   /// * [enableBnpl] - функционал оплаты частями
   static Future<bool> initSberPay({
+    required String apiKey,
+    required String merchantLogin,
     required SberPayEnv env,
     bool? enableBnpl,
   }) async =>
       _platform.initSberPay(
         SberPayInitConfig(
+          apiKey: apiKey,
+          merchantLogin: merchantLogin,
           env: env,
           enableBnpl: enableBnpl,
         ),
@@ -32,21 +35,14 @@ class SberPayPlugin {
   /// Зависит от переданного аргумента *env* при инициализации через метод
   /// [initSberPay] (см. комментарий к методу).
   ///
-  /// Метод инициализации синхронный, хотя в нем выполняется асинхронная
-  /// операция, которая неизвестно когда выполнится. Поэтому сначала нужно
-  /// вызвать [initSberPay], подождать, к примеру, секунды 2 и только потом
-  /// обращаться к этому методу (см. пример в example/lib/main.dart).
-  ///
   /// Если у пользователя нет установленного сбера в режимах
   /// [SberPayEnv.sandboxRealBankApp], [SberPayEnv.prod] - вернет false.
   static Future<bool> isReadyForSPaySdk() async =>
       _platform.isReadyForSPaySdk();
 
   /// Метод оплаты через SberPay SDK.
-  /// * [apiKey] - ключ, выдаваемый по договору, либо создаваемый в личном
-  /// кабинете;
-  /// * [merchantLogin] - логин, выдаваемый по договору, либо создаваемый в
-  /// личном кабинете;
+  /// * [apiKey] - ключ, выдаваемый по договору (необязательно, если передан в init)
+  /// * [merchantLogin] - логин, выдаваемый по договору (необязательно, если передан в init)
   /// * [bankInvoiceId] - параметр, который получаем после запроса для
   /// регистрации заказа в шлюзе Сбера.
   /// * [redirectUri] - диплинк для перехода обратно в приложение после открытия
@@ -56,11 +52,11 @@ class SberPayPlugin {
   ///
   /// Возвращает статус оплаты [SberPayPaymentStatus]
   static Future<SberPayPaymentStatus> payWithBankInvoiceId({
-    required String apiKey,
-    required String merchantLogin,
     required String bankInvoiceId,
     required String redirectUri,
     required String orderNumber,
+    String? apiKey,
+    String? merchantLogin,
   }) async =>
       _platform.payWithBankInvoiceId(
         SberPayPaymentConfig(
