@@ -1,33 +1,68 @@
-// Autogenerated from Pigeon (v16.0.0), do not edit directly.
+// Autogenerated from Pigeon (v26.1.4), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
+@file:Suppress("UNCHECKED_CAST", "ArrayInDataClass")
 
 
 import android.util.Log
 import io.flutter.plugin.common.BasicMessageChannel
 import io.flutter.plugin.common.BinaryMessenger
+import io.flutter.plugin.common.EventChannel
 import io.flutter.plugin.common.MessageCodec
+import io.flutter.plugin.common.StandardMethodCodec
 import io.flutter.plugin.common.StandardMessageCodec
 import java.io.ByteArrayOutputStream
 import java.nio.ByteBuffer
+private object MessagesPigeonUtils {
 
-private fun wrapResult(result: Any?): List<Any?> {
-  return listOf(result)
-}
-
-private fun wrapError(exception: Throwable): List<Any?> {
-  if (exception is FlutterError) {
-    return listOf(
-      exception.code,
-      exception.message,
-      exception.details
-    )
-  } else {
-    return listOf(
-      exception.javaClass.simpleName,
-      exception.toString(),
-      "Cause: " + exception.cause + ", Stacktrace: " + Log.getStackTraceString(exception)
-    )
+  fun wrapResult(result: Any?): List<Any?> {
+    return listOf(result)
   }
+
+  fun wrapError(exception: Throwable): List<Any?> {
+    return if (exception is FlutterError) {
+      listOf(
+        exception.code,
+        exception.message,
+        exception.details
+      )
+    } else {
+      listOf(
+        exception.javaClass.simpleName,
+        exception.toString(),
+        "Cause: " + exception.cause + ", Stacktrace: " + Log.getStackTraceString(exception)
+      )
+    }
+  }
+  fun deepEquals(a: Any?, b: Any?): Boolean {
+    if (a is ByteArray && b is ByteArray) {
+        return a.contentEquals(b)
+    }
+    if (a is IntArray && b is IntArray) {
+        return a.contentEquals(b)
+    }
+    if (a is LongArray && b is LongArray) {
+        return a.contentEquals(b)
+    }
+    if (a is DoubleArray && b is DoubleArray) {
+        return a.contentEquals(b)
+    }
+    if (a is Array<*> && b is Array<*>) {
+      return a.size == b.size &&
+          a.indices.all{ deepEquals(a[it], b[it]) }
+    }
+    if (a is List<*> && b is List<*>) {
+      return a.size == b.size &&
+          a.indices.all{ deepEquals(a[it], b[it]) }
+    }
+    if (a is Map<*, *> && b is Map<*, *>) {
+      return a.size == b.size && a.all {
+          (b as Map<Any?, Any?>).contains(it.key) &&
+          deepEquals(it.value, b[it.key])
+      }
+    }
+    return a == b
+  }
+      
 }
 
 /**
@@ -55,14 +90,14 @@ enum class SberPayApiEnv(val raw: Int) {
    *
    * Позволяет протестировать оплату как в [prod], но с тестовыми данными.
    */
-  SANDBOXREALBANKAPP(1),
+  SANDBOX_REAL_BANK_APP(1),
   /**
    * Режим песочницы без перехода в банк.
    *
    * При авторизации пользователя не осуществляется переход в приложение
    * Сбербанка.
    */
-  SANDBOXWITHOUTBANKAPP(2);
+  SANDBOX_WITHOUT_BANK_APP(2);
 
   companion object {
     fun ofRaw(raw: Int): SberPayApiEnv? {
@@ -95,26 +130,43 @@ enum class SberPayApiPaymentStatus(val raw: Int) {
  * Generated class from Pigeon that represents data sent in messages.
  */
 data class InitConfig (
+  /** Ключ, выдаваемый по договору, либо создаваемый в личном кабинете */
+  val apiKey: String,
+  /** Логин, выдаваемый по договору, либо создаваемый в личном кабинете */
+  val merchantLogin: String,
   /** Среда запуска */
   val env: SberPayApiEnv,
   /** Использование функционала оплаты частями */
   val enableBnpl: Boolean? = null
-
-) {
+)
+ {
   companion object {
-    @Suppress("UNCHECKED_CAST")
-    fun fromList(list: List<Any?>): InitConfig {
-      val env = SberPayApiEnv.ofRaw(list[0] as Int)!!
-      val enableBnpl = list[1] as Boolean?
-      return InitConfig(env, enableBnpl)
+    fun fromList(pigeonVar_list: List<Any?>): InitConfig {
+      val apiKey = pigeonVar_list[0] as String
+      val merchantLogin = pigeonVar_list[1] as String
+      val env = pigeonVar_list[2] as SberPayApiEnv
+      val enableBnpl = pigeonVar_list[3] as Boolean?
+      return InitConfig(apiKey, merchantLogin, env, enableBnpl)
     }
   }
   fun toList(): List<Any?> {
-    return listOf<Any?>(
-      env.raw,
+    return listOf(
+      apiKey,
+      merchantLogin,
+      env,
       enableBnpl,
     )
   }
+  override fun equals(other: Any?): Boolean {
+    if (other !is InitConfig) {
+      return false
+    }
+    if (this === other) {
+      return true
+    }
+    return MessagesPigeonUtils.deepEquals(toList(), other.toList())  }
+
+  override fun hashCode(): Int = toList().hashCode()
 }
 
 /**
@@ -123,46 +175,55 @@ data class InitConfig (
  * Generated class from Pigeon that represents data sent in messages.
  */
 data class PayConfig (
-  /** Ключ, выдаваемый по договору, либо создаваемый в личном кабинете */
-  val apiKey: String,
-  /** Логин, выдаваемый по договору, либо создаваемый в личном кабинете */
-  val merchantLogin: String,
   /** Уникальный идентификатор заказа, сгенерированный Банком */
   val bankInvoiceId: String,
   /** Номер заказа */
   val orderNumber: String
-
-) {
+)
+ {
   companion object {
-    @Suppress("UNCHECKED_CAST")
-    fun fromList(list: List<Any?>): PayConfig {
-      val apiKey = list[0] as String
-      val merchantLogin = list[1] as String
-      val bankInvoiceId = list[2] as String
-      val orderNumber = list[3] as String
-      return PayConfig(apiKey, merchantLogin, bankInvoiceId, orderNumber)
+    fun fromList(pigeonVar_list: List<Any?>): PayConfig {
+      val bankInvoiceId = pigeonVar_list[0] as String
+      val orderNumber = pigeonVar_list[1] as String
+      return PayConfig(bankInvoiceId, orderNumber)
     }
   }
   fun toList(): List<Any?> {
-    return listOf<Any?>(
-      apiKey,
-      merchantLogin,
+    return listOf(
       bankInvoiceId,
       orderNumber,
     )
   }
-}
+  override fun equals(other: Any?): Boolean {
+    if (other !is PayConfig) {
+      return false
+    }
+    if (this === other) {
+      return true
+    }
+    return MessagesPigeonUtils.deepEquals(toList(), other.toList())  }
 
-@Suppress("UNCHECKED_CAST")
-private object SberPayApiCodec : StandardMessageCodec() {
+  override fun hashCode(): Int = toList().hashCode()
+}
+private open class MessagesPigeonCodec : StandardMessageCodec() {
   override fun readValueOfType(type: Byte, buffer: ByteBuffer): Any? {
     return when (type) {
-      128.toByte() -> {
+      129.toByte() -> {
+        return (readValue(buffer) as Long?)?.let {
+          SberPayApiEnv.ofRaw(it.toInt())
+        }
+      }
+      130.toByte() -> {
+        return (readValue(buffer) as Long?)?.let {
+          SberPayApiPaymentStatus.ofRaw(it.toInt())
+        }
+      }
+      131.toByte() -> {
         return (readValue(buffer) as? List<Any?>)?.let {
           InitConfig.fromList(it)
         }
       }
-      129.toByte() -> {
+      132.toByte() -> {
         return (readValue(buffer) as? List<Any?>)?.let {
           PayConfig.fromList(it)
         }
@@ -172,12 +233,20 @@ private object SberPayApiCodec : StandardMessageCodec() {
   }
   override fun writeValue(stream: ByteArrayOutputStream, value: Any?)   {
     when (value) {
+      is SberPayApiEnv -> {
+        stream.write(129)
+        writeValue(stream, value.raw.toLong())
+      }
+      is SberPayApiPaymentStatus -> {
+        stream.write(130)
+        writeValue(stream, value.raw.toLong())
+      }
       is InitConfig -> {
-        stream.write(128)
+        stream.write(131)
         writeValue(stream, value.toList())
       }
       is PayConfig -> {
-        stream.write(129)
+        stream.write(132)
         writeValue(stream, value.toList())
       }
       else -> super.writeValue(stream, value)
@@ -185,47 +254,50 @@ private object SberPayApiCodec : StandardMessageCodec() {
   }
 }
 
+
 /** Generated interface from Pigeon that represents a handler of messages from Flutter. */
 interface SberPayApi {
-  fun initSberPay(config: InitConfig): Boolean
+  fun initSberPay(config: InitConfig, callback: (Result<Boolean>) -> Unit)
   fun isReadyForSPaySdk(): Boolean
   fun payWithBankInvoiceId(config: PayConfig, callback: (Result<SberPayApiPaymentStatus>) -> Unit)
 
   companion object {
     /** The codec used by SberPayApi. */
     val codec: MessageCodec<Any?> by lazy {
-      SberPayApiCodec
+      MessagesPigeonCodec()
     }
     /** Sets up an instance of `SberPayApi` to handle messages through the `binaryMessenger`. */
-    @Suppress("UNCHECKED_CAST")
-    fun setUp(binaryMessenger: BinaryMessenger, api: SberPayApi?) {
+    @JvmOverloads
+    fun setUp(binaryMessenger: BinaryMessenger, api: SberPayApi?, messageChannelSuffix: String = "") {
+      val separatedMessageChannelSuffix = if (messageChannelSuffix.isNotEmpty()) ".$messageChannelSuffix" else ""
       run {
-        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.sber_pay_android.SberPayApi.initSberPay", codec)
+        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.sber_pay_android.SberPayApi.initSberPay$separatedMessageChannelSuffix", codec)
         if (api != null) {
           channel.setMessageHandler { message, reply ->
             val args = message as List<Any?>
             val configArg = args[0] as InitConfig
-            var wrapped: List<Any?>
-            try {
-              wrapped = listOf<Any?>(api.initSberPay(configArg))
-            } catch (exception: Throwable) {
-              wrapped = wrapError(exception)
+            api.initSberPay(configArg) { result: Result<Boolean> ->
+              val error = result.exceptionOrNull()
+              if (error != null) {
+                reply.reply(MessagesPigeonUtils.wrapError(error))
+              } else {
+                val data = result.getOrNull()
+                reply.reply(MessagesPigeonUtils.wrapResult(data))
+              }
             }
-            reply.reply(wrapped)
           }
         } else {
           channel.setMessageHandler(null)
         }
       }
       run {
-        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.sber_pay_android.SberPayApi.isReadyForSPaySdk", codec)
+        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.sber_pay_android.SberPayApi.isReadyForSPaySdk$separatedMessageChannelSuffix", codec)
         if (api != null) {
           channel.setMessageHandler { _, reply ->
-            var wrapped: List<Any?>
-            try {
-              wrapped = listOf<Any?>(api.isReadyForSPaySdk())
+            val wrapped: List<Any?> = try {
+              listOf(api.isReadyForSPaySdk())
             } catch (exception: Throwable) {
-              wrapped = wrapError(exception)
+              MessagesPigeonUtils.wrapError(exception)
             }
             reply.reply(wrapped)
           }
@@ -234,7 +306,7 @@ interface SberPayApi {
         }
       }
       run {
-        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.sber_pay_android.SberPayApi.payWithBankInvoiceId", codec)
+        val channel = BasicMessageChannel<Any?>(binaryMessenger, "dev.flutter.pigeon.sber_pay_android.SberPayApi.payWithBankInvoiceId$separatedMessageChannelSuffix", codec)
         if (api != null) {
           channel.setMessageHandler { message, reply ->
             val args = message as List<Any?>
@@ -242,10 +314,10 @@ interface SberPayApi {
             api.payWithBankInvoiceId(configArg) { result: Result<SberPayApiPaymentStatus> ->
               val error = result.exceptionOrNull()
               if (error != null) {
-                reply.reply(wrapError(error))
+                reply.reply(MessagesPigeonUtils.wrapError(error))
               } else {
                 val data = result.getOrNull()
-                reply.reply(wrapResult(data!!.raw))
+                reply.reply(MessagesPigeonUtils.wrapResult(data))
               }
             }
           }
