@@ -1,6 +1,6 @@
-// Autogenerated from Pigeon (v16.0.0), do not edit directly.
+// Autogenerated from Pigeon (v26.1.4), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
-// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types, unused_shown_name, unnecessary_import, no_leading_underscores_for_local_identifiers
+// ignore_for_file: public_member_api_docs, non_constant_identifier_names, avoid_as, unused_import, unnecessary_parenthesis, prefer_null_aware_operators, omit_local_variable_types, omit_obvious_local_variable_types, unused_shown_name, unnecessary_import, no_leading_underscores_for_local_identifiers
 
 import 'dart:async';
 import 'dart:typed_data' show Float64List, Int32List, Int64List, Uint8List;
@@ -15,16 +15,33 @@ PlatformException _createConnectionError(String channelName) {
   );
 }
 
+bool _deepEquals(Object? a, Object? b) {
+  if (a is List && b is List) {
+    return a.length == b.length &&
+        a.indexed
+            .every(((int, dynamic) item) => _deepEquals(item.$2, b[item.$1]));
+  }
+  if (a is Map && b is Map) {
+    return a.length == b.length &&
+        a.entries.every((MapEntry<Object?, Object?> entry) =>
+            (b as Map<Object?, Object?>).containsKey(entry.key) &&
+            _deepEquals(entry.value, b[entry.key]));
+  }
+  return a == b;
+}
+
 /// Тип инициализации сервисов Сбербанка
 enum SberPayApiEnv {
   /// Продуктовый режим.
   ///
   /// Для авторизации пользователя происходит редирект в приложение Сбербанка.
   prod,
+
   /// Режим песочницы.
   ///
   /// Позволяет протестировать оплату как в [prod], но с тестовыми данными.
   sandboxRealBankApp,
+
   /// Режим песочницы без перехода в банк.
   ///
   /// При авторизации пользователя не осуществляется переход в приложение
@@ -36,10 +53,13 @@ enum SberPayApiEnv {
 enum SberPayApiPaymentStatus {
   /// Успешный результат
   success,
+
   /// Необходимо проверить статус оплаты
   processing,
+
   /// Пользователь отменил оплату
   cancel,
+
   /// Неизвестный тип
   unknown,
 }
@@ -47,39 +67,10 @@ enum SberPayApiPaymentStatus {
 /// Конфигурация инициализации
 class InitConfig {
   InitConfig({
-    required this.env,
-    this.enableBnpl,
-  });
-
-  /// Среда запуска
-  SberPayApiEnv env;
-
-  /// Использование функционала оплаты частями
-  bool? enableBnpl;
-
-  Object encode() {
-    return <Object?>[
-      env.index,
-      enableBnpl,
-    ];
-  }
-
-  static InitConfig decode(Object result) {
-    result as List<Object?>;
-    return InitConfig(
-      env: SberPayApiEnv.values[result[0]! as int],
-      enableBnpl: result[1] as bool?,
-    );
-  }
-}
-
-/// Конфигурация оплаты
-class PayConfig {
-  PayConfig({
     required this.apiKey,
     required this.merchantLogin,
-    required this.bankInvoiceId,
-    required this.orderNumber,
+    required this.env,
+    this.enableBnpl,
   });
 
   /// Ключ, выдаваемый по договору, либо создаваемый в личном кабинете
@@ -88,41 +79,119 @@ class PayConfig {
   /// Логин, выдаваемый по договору, либо создаваемый в личном кабинете
   String merchantLogin;
 
+  /// Среда запуска
+  SberPayApiEnv env;
+
+  /// Использование функционала оплаты частями
+  bool? enableBnpl;
+
+  List<Object?> _toList() {
+    return <Object?>[
+      apiKey,
+      merchantLogin,
+      env,
+      enableBnpl,
+    ];
+  }
+
+  Object encode() {
+    return _toList();
+  }
+
+  static InitConfig decode(Object result) {
+    result as List<Object?>;
+    return InitConfig(
+      apiKey: result[0]! as String,
+      merchantLogin: result[1]! as String,
+      env: result[2]! as SberPayApiEnv,
+      enableBnpl: result[3] as bool?,
+    );
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  bool operator ==(Object other) {
+    if (other is! InitConfig || other.runtimeType != runtimeType) {
+      return false;
+    }
+    if (identical(this, other)) {
+      return true;
+    }
+    return _deepEquals(encode(), other.encode());
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  int get hashCode => Object.hashAll(_toList());
+}
+
+/// Конфигурация оплаты
+class PayConfig {
+  PayConfig({
+    required this.bankInvoiceId,
+    required this.orderNumber,
+  });
+
   /// Уникальный идентификатор заказа, сгенерированный Банком
   String bankInvoiceId;
 
   /// Номер заказа
   String orderNumber;
 
-  Object encode() {
+  List<Object?> _toList() {
     return <Object?>[
-      apiKey,
-      merchantLogin,
       bankInvoiceId,
       orderNumber,
     ];
   }
 
+  Object encode() {
+    return _toList();
+  }
+
   static PayConfig decode(Object result) {
     result as List<Object?>;
     return PayConfig(
-      apiKey: result[0]! as String,
-      merchantLogin: result[1]! as String,
-      bankInvoiceId: result[2]! as String,
-      orderNumber: result[3]! as String,
+      bankInvoiceId: result[0]! as String,
+      orderNumber: result[1]! as String,
     );
   }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  bool operator ==(Object other) {
+    if (other is! PayConfig || other.runtimeType != runtimeType) {
+      return false;
+    }
+    if (identical(this, other)) {
+      return true;
+    }
+    return _deepEquals(encode(), other.encode());
+  }
+
+  @override
+  // ignore: avoid_equals_and_hash_code_on_mutable_classes
+  int get hashCode => Object.hashAll(_toList());
 }
 
-class _SberPayApiCodec extends StandardMessageCodec {
-  const _SberPayApiCodec();
+class _PigeonCodec extends StandardMessageCodec {
+  const _PigeonCodec();
   @override
   void writeValue(WriteBuffer buffer, Object? value) {
-    if (value is InitConfig) {
-      buffer.putUint8(128);
+    if (value is int) {
+      buffer.putUint8(4);
+      buffer.putInt64(value);
+    } else if (value is SberPayApiEnv) {
+      buffer.putUint8(129);
+      writeValue(buffer, value.index);
+    } else if (value is SberPayApiPaymentStatus) {
+      buffer.putUint8(130);
+      writeValue(buffer, value.index);
+    } else if (value is InitConfig) {
+      buffer.putUint8(131);
       writeValue(buffer, value.encode());
     } else if (value is PayConfig) {
-      buffer.putUint8(129);
+      buffer.putUint8(132);
       writeValue(buffer, value.encode());
     } else {
       super.writeValue(buffer, value);
@@ -132,9 +201,15 @@ class _SberPayApiCodec extends StandardMessageCodec {
   @override
   Object? readValueOfType(int type, ReadBuffer buffer) {
     switch (type) {
-      case 128: 
+      case 129:
+        final value = readValue(buffer) as int?;
+        return value == null ? null : SberPayApiEnv.values[value];
+      case 130:
+        final value = readValue(buffer) as int?;
+        return value == null ? null : SberPayApiPaymentStatus.values[value];
+      case 131:
         return InitConfig.decode(readValue(buffer)!);
-      case 129: 
+      case 132:
         return PayConfig.decode(readValue(buffer)!);
       default:
         return super.readValueOfType(type, buffer);
@@ -146,90 +221,100 @@ class SberPayApi {
   /// Constructor for [SberPayApi].  The [binaryMessenger] named argument is
   /// available for dependency injection.  If it is left null, the default
   /// BinaryMessenger will be used which routes to the host platform.
-  SberPayApi({BinaryMessenger? binaryMessenger})
-      : __pigeon_binaryMessenger = binaryMessenger;
-  final BinaryMessenger? __pigeon_binaryMessenger;
+  SberPayApi(
+      {BinaryMessenger? binaryMessenger, String messageChannelSuffix = ''})
+      : pigeonVar_binaryMessenger = binaryMessenger,
+        pigeonVar_messageChannelSuffix =
+            messageChannelSuffix.isNotEmpty ? '.$messageChannelSuffix' : '';
+  final BinaryMessenger? pigeonVar_binaryMessenger;
 
-  static const MessageCodec<Object?> pigeonChannelCodec = _SberPayApiCodec();
+  static const MessageCodec<Object?> pigeonChannelCodec = _PigeonCodec();
+
+  final String pigeonVar_messageChannelSuffix;
 
   Future<bool> initSberPay(InitConfig config) async {
-    const String __pigeon_channelName = 'dev.flutter.pigeon.sber_pay_android.SberPayApi.initSberPay';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-      __pigeon_channelName,
+    final pigeonVar_channelName =
+        'dev.flutter.pigeon.sber_pay_android.SberPayApi.initSberPay$pigeonVar_messageChannelSuffix';
+    final pigeonVar_channel = BasicMessageChannel<Object?>(
+      pigeonVar_channelName,
       pigeonChannelCodec,
-      binaryMessenger: __pigeon_binaryMessenger,
+      binaryMessenger: pigeonVar_binaryMessenger,
     );
-    final List<Object?>? __pigeon_replyList =
-        await __pigeon_channel.send(<Object?>[config]) as List<Object?>?;
-    if (__pigeon_replyList == null) {
-      throw _createConnectionError(__pigeon_channelName);
-    } else if (__pigeon_replyList.length > 1) {
+    final Future<Object?> pigeonVar_sendFuture =
+        pigeonVar_channel.send(<Object?>[config]);
+    final pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    if (pigeonVar_replyList == null) {
+      throw _createConnectionError(pigeonVar_channelName);
+    } else if (pigeonVar_replyList.length > 1) {
       throw PlatformException(
-        code: __pigeon_replyList[0]! as String,
-        message: __pigeon_replyList[1] as String?,
-        details: __pigeon_replyList[2],
+        code: pigeonVar_replyList[0]! as String,
+        message: pigeonVar_replyList[1] as String?,
+        details: pigeonVar_replyList[2],
       );
-    } else if (__pigeon_replyList[0] == null) {
+    } else if (pigeonVar_replyList[0] == null) {
       throw PlatformException(
         code: 'null-error',
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (__pigeon_replyList[0] as bool?)!;
+      return (pigeonVar_replyList[0] as bool?)!;
     }
   }
 
   Future<bool> isReadyForSPaySdk() async {
-    const String __pigeon_channelName = 'dev.flutter.pigeon.sber_pay_android.SberPayApi.isReadyForSPaySdk';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-      __pigeon_channelName,
+    final pigeonVar_channelName =
+        'dev.flutter.pigeon.sber_pay_android.SberPayApi.isReadyForSPaySdk$pigeonVar_messageChannelSuffix';
+    final pigeonVar_channel = BasicMessageChannel<Object?>(
+      pigeonVar_channelName,
       pigeonChannelCodec,
-      binaryMessenger: __pigeon_binaryMessenger,
+      binaryMessenger: pigeonVar_binaryMessenger,
     );
-    final List<Object?>? __pigeon_replyList =
-        await __pigeon_channel.send(null) as List<Object?>?;
-    if (__pigeon_replyList == null) {
-      throw _createConnectionError(__pigeon_channelName);
-    } else if (__pigeon_replyList.length > 1) {
+    final Future<Object?> pigeonVar_sendFuture = pigeonVar_channel.send(null);
+    final pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    if (pigeonVar_replyList == null) {
+      throw _createConnectionError(pigeonVar_channelName);
+    } else if (pigeonVar_replyList.length > 1) {
       throw PlatformException(
-        code: __pigeon_replyList[0]! as String,
-        message: __pigeon_replyList[1] as String?,
-        details: __pigeon_replyList[2],
+        code: pigeonVar_replyList[0]! as String,
+        message: pigeonVar_replyList[1] as String?,
+        details: pigeonVar_replyList[2],
       );
-    } else if (__pigeon_replyList[0] == null) {
+    } else if (pigeonVar_replyList[0] == null) {
       throw PlatformException(
         code: 'null-error',
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return (__pigeon_replyList[0] as bool?)!;
+      return (pigeonVar_replyList[0] as bool?)!;
     }
   }
 
   Future<SberPayApiPaymentStatus> payWithBankInvoiceId(PayConfig config) async {
-    const String __pigeon_channelName = 'dev.flutter.pigeon.sber_pay_android.SberPayApi.payWithBankInvoiceId';
-    final BasicMessageChannel<Object?> __pigeon_channel = BasicMessageChannel<Object?>(
-      __pigeon_channelName,
+    final pigeonVar_channelName =
+        'dev.flutter.pigeon.sber_pay_android.SberPayApi.payWithBankInvoiceId$pigeonVar_messageChannelSuffix';
+    final pigeonVar_channel = BasicMessageChannel<Object?>(
+      pigeonVar_channelName,
       pigeonChannelCodec,
-      binaryMessenger: __pigeon_binaryMessenger,
+      binaryMessenger: pigeonVar_binaryMessenger,
     );
-    final List<Object?>? __pigeon_replyList =
-        await __pigeon_channel.send(<Object?>[config]) as List<Object?>?;
-    if (__pigeon_replyList == null) {
-      throw _createConnectionError(__pigeon_channelName);
-    } else if (__pigeon_replyList.length > 1) {
+    final Future<Object?> pigeonVar_sendFuture =
+        pigeonVar_channel.send(<Object?>[config]);
+    final pigeonVar_replyList = await pigeonVar_sendFuture as List<Object?>?;
+    if (pigeonVar_replyList == null) {
+      throw _createConnectionError(pigeonVar_channelName);
+    } else if (pigeonVar_replyList.length > 1) {
       throw PlatformException(
-        code: __pigeon_replyList[0]! as String,
-        message: __pigeon_replyList[1] as String?,
-        details: __pigeon_replyList[2],
+        code: pigeonVar_replyList[0]! as String,
+        message: pigeonVar_replyList[1] as String?,
+        details: pigeonVar_replyList[2],
       );
-    } else if (__pigeon_replyList[0] == null) {
+    } else if (pigeonVar_replyList[0] == null) {
       throw PlatformException(
         code: 'null-error',
         message: 'Host platform returned null value for non-null return value.',
       );
     } else {
-      return SberPayApiPaymentStatus.values[__pigeon_replyList[0]! as int];
+      return (pigeonVar_replyList[0] as SberPayApiPaymentStatus?)!;
     }
   }
 }
