@@ -44,9 +44,17 @@ enum SberPayApiPaymentStatus {
 /// Конфигурация инициализации
 class InitConfig {
   const InitConfig({
+    required this.apiKey,
+    required this.merchantLogin,
     required this.env,
     required this.enableBnpl,
   });
+
+  /// Ключ, выдаваемый по договору, либо создаваемый в личном кабинете
+  final String apiKey;
+
+  /// Логин, выдаваемый по договору, либо создаваемый в личном кабинете
+  final String merchantLogin;
 
   /// Среда запуска
   final SberPayApiEnv env;
@@ -58,17 +66,9 @@ class InitConfig {
 /// Конфигурация оплаты
 class PayConfig {
   const PayConfig({
-    required this.apiKey,
-    required this.merchantLogin,
     required this.bankInvoiceId,
     required this.orderNumber,
   });
-
-  /// Ключ, выдаваемый по договору, либо создаваемый в личном кабинете
-  final String apiKey;
-
-  /// Логин, выдаваемый по договору, либо создаваемый в личном кабинете
-  final String merchantLogin;
 
   /// Уникальный идентификатор заказа, сгенерированный Банком
   final String bankInvoiceId;
@@ -79,6 +79,7 @@ class PayConfig {
 
 @HostApi()
 abstract class SberPayApi {
+  @async
   bool initSberPay(InitConfig config);
 
   bool isReadyForSPaySdk();
