@@ -1,31 +1,57 @@
-// Autogenerated from Pigeon (v16.0.0), do not edit directly.
+// Autogenerated from Pigeon (v26.1.4), do not edit directly.
 // See also: https://pub.dev/packages/pigeon
 
 import Foundation
+
 #if os(iOS)
-import Flutter
+  import Flutter
 #elseif os(macOS)
-import FlutterMacOS
+  import FlutterMacOS
 #else
-#error("Unsupported platform.")
+  #error("Unsupported platform.")
 #endif
+
+/// Error class for passing custom error details to Dart side.
+final class PigeonError: Error {
+  let code: String
+  let message: String?
+  let details: Sendable?
+
+  init(code: String, message: String?, details: Sendable?) {
+    self.code = code
+    self.message = message
+    self.details = details
+  }
+
+  var localizedDescription: String {
+    return
+      "PigeonError(code: \(code), message: \(message ?? "<nil>"), details: \(details ?? "<nil>")"
+  }
+}
 
 private func wrapResult(_ result: Any?) -> [Any?] {
   return [result]
 }
 
 private func wrapError(_ error: Any) -> [Any?] {
+  if let pigeonError = error as? PigeonError {
+    return [
+      pigeonError.code,
+      pigeonError.message,
+      pigeonError.details,
+    ]
+  }
   if let flutterError = error as? FlutterError {
     return [
       flutterError.code,
       flutterError.message,
-      flutterError.details
+      flutterError.details,
     ]
   }
   return [
     "\(error)",
     "\(type(of: error))",
-    "Stacktrace: \(Thread.callStackSymbols)"
+    "Stacktrace: \(Thread.callStackSymbols)",
   ]
 }
 
@@ -37,6 +63,70 @@ private func nilOrValue<T>(_ value: Any?) -> T? {
   if value is NSNull { return nil }
   return value as! T?
 }
+
+func deepEqualsMessages(_ lhs: Any?, _ rhs: Any?) -> Bool {
+  let cleanLhs = nilOrValue(lhs) as Any?
+  let cleanRhs = nilOrValue(rhs) as Any?
+  switch (cleanLhs, cleanRhs) {
+  case (nil, nil):
+    return true
+
+  case (nil, _), (_, nil):
+    return false
+
+  case is (Void, Void):
+    return true
+
+  case let (cleanLhsHashable, cleanRhsHashable) as (AnyHashable, AnyHashable):
+    return cleanLhsHashable == cleanRhsHashable
+
+  case let (cleanLhsArray, cleanRhsArray) as ([Any?], [Any?]):
+    guard cleanLhsArray.count == cleanRhsArray.count else { return false }
+    for (index, element) in cleanLhsArray.enumerated() {
+      if !deepEqualsMessages(element, cleanRhsArray[index]) {
+        return false
+      }
+    }
+    return true
+
+  case let (cleanLhsDictionary, cleanRhsDictionary) as ([AnyHashable: Any?], [AnyHashable: Any?]):
+    guard cleanLhsDictionary.count == cleanRhsDictionary.count else { return false }
+    for (key, cleanLhsValue) in cleanLhsDictionary {
+      guard cleanRhsDictionary.index(forKey: key) != nil else { return false }
+      if !deepEqualsMessages(cleanLhsValue, cleanRhsDictionary[key]!) {
+        return false
+      }
+    }
+    return true
+
+  default:
+    // Any other type shouldn't be able to be used with pigeon. File an issue if you find this to be untrue.
+    return false
+  }
+}
+
+func deepHashMessages(value: Any?, hasher: inout Hasher) {
+  if let valueList = value as? [AnyHashable] {
+     for item in valueList { deepHashMessages(value: item, hasher: &hasher) }
+     return
+  }
+
+  if let valueDict = value as? [AnyHashable: AnyHashable] {
+    for key in valueDict.keys { 
+      hasher.combine(key)
+      deepHashMessages(value: valueDict[key]!, hasher: &hasher)
+    }
+    return
+  }
+
+  if let hashableValue = value as? AnyHashable {
+    hasher.combine(hashableValue.hashValue)
+  }
+
+  return hasher.combine(String(describing: value))
+}
+
+    
 
 /// Тип инициализации сервисов Сбербанка
 enum SberPayApiEnv: Int {
@@ -70,33 +160,50 @@ enum SberPayApiPaymentStatus: Int {
 /// Конфигурация инициализации
 ///
 /// Generated class from Pigeon that represents data sent in messages.
-struct InitConfig {
+struct InitConfig: Hashable {
+  /// Ключ, выдаваемый по договору, либо создаваемый в личном кабинете
+  var apiKey: String
+  /// Логин, выдаваемый по договору, либо создаваемый в личном кабинете
+  var merchantLogin: String
   /// Среда запуска
   var env: SberPayApiEnv
   /// Использование функционала оплаты частями
   var enableBnpl: Bool? = nil
 
-  static func fromList(_ list: [Any?]) -> InitConfig? {
-    let env = SberPayApiEnv(rawValue: list[0] as! Int)!
-    let enableBnpl: Bool? = nilOrValue(list[1])
+
+  // swift-format-ignore: AlwaysUseLowerCamelCase
+  static func fromList(_ pigeonVar_list: [Any?]) -> InitConfig? {
+    let apiKey = pigeonVar_list[0] as! String
+    let merchantLogin = pigeonVar_list[1] as! String
+    let env = pigeonVar_list[2] as! SberPayApiEnv
+    let enableBnpl: Bool? = nilOrValue(pigeonVar_list[3])
 
     return InitConfig(
+      apiKey: apiKey,
+      merchantLogin: merchantLogin,
       env: env,
       enableBnpl: enableBnpl
     )
   }
   func toList() -> [Any?] {
     return [
-      env.rawValue,
+      apiKey,
+      merchantLogin,
+      env,
       enableBnpl,
     ]
+  }
+  static func == (lhs: InitConfig, rhs: InitConfig) -> Bool {
+    return deepEqualsMessages(lhs.toList(), rhs.toList())  }
+  func hash(into hasher: inout Hasher) {
+    deepHashMessages(value: toList(), hasher: &hasher)
   }
 }
 
 /// Конфигурация оплаты
 ///
 /// Generated class from Pigeon that represents data sent in messages.
-struct PayConfig {
+struct PayConfig: Hashable {
   /// Ключ, выдаваемый по договору, либо создаваемый в личном кабинете
   var apiKey: String
   /// Логин, выдаваемый по договору, либо создаваемый в личном кабинете
@@ -108,12 +215,14 @@ struct PayConfig {
   /// Номер заказа
   var orderNumber: String
 
-  static func fromList(_ list: [Any?]) -> PayConfig? {
-    let apiKey = list[0] as! String
-    let merchantLogin = list[1] as! String
-    let bankInvoiceId = list[2] as! String
-    let redirectUri = list[3] as! String
-    let orderNumber = list[4] as! String
+
+  // swift-format-ignore: AlwaysUseLowerCamelCase
+  static func fromList(_ pigeonVar_list: [Any?]) -> PayConfig? {
+    let apiKey = pigeonVar_list[0] as! String
+    let merchantLogin = pigeonVar_list[1] as! String
+    let bankInvoiceId = pigeonVar_list[2] as! String
+    let redirectUri = pigeonVar_list[3] as! String
+    let orderNumber = pigeonVar_list[4] as! String
 
     return PayConfig(
       apiKey: apiKey,
@@ -132,28 +241,51 @@ struct PayConfig {
       orderNumber,
     ]
   }
+  static func == (lhs: PayConfig, rhs: PayConfig) -> Bool {
+    return deepEqualsMessages(lhs.toList(), rhs.toList())  }
+  func hash(into hasher: inout Hasher) {
+    deepHashMessages(value: toList(), hasher: &hasher)
+  }
 }
 
-private class SberPayApiCodecReader: FlutterStandardReader {
+private class MessagesPigeonCodecReader: FlutterStandardReader {
   override func readValue(ofType type: UInt8) -> Any? {
     switch type {
-      case 128:
-        return InitConfig.fromList(self.readValue() as! [Any?])
-      case 129:
-        return PayConfig.fromList(self.readValue() as! [Any?])
-      default:
-        return super.readValue(ofType: type)
+    case 129:
+      let enumResultAsInt: Int? = nilOrValue(self.readValue() as! Int?)
+      if let enumResultAsInt = enumResultAsInt {
+        return SberPayApiEnv(rawValue: enumResultAsInt)
+      }
+      return nil
+    case 130:
+      let enumResultAsInt: Int? = nilOrValue(self.readValue() as! Int?)
+      if let enumResultAsInt = enumResultAsInt {
+        return SberPayApiPaymentStatus(rawValue: enumResultAsInt)
+      }
+      return nil
+    case 131:
+      return InitConfig.fromList(self.readValue() as! [Any?])
+    case 132:
+      return PayConfig.fromList(self.readValue() as! [Any?])
+    default:
+      return super.readValue(ofType: type)
     }
   }
 }
 
-private class SberPayApiCodecWriter: FlutterStandardWriter {
+private class MessagesPigeonCodecWriter: FlutterStandardWriter {
   override func writeValue(_ value: Any) {
-    if let value = value as? InitConfig {
-      super.writeByte(128)
+    if let value = value as? SberPayApiEnv {
+      super.writeByte(129)
+      super.writeValue(value.rawValue)
+    } else if let value = value as? SberPayApiPaymentStatus {
+      super.writeByte(130)
+      super.writeValue(value.rawValue)
+    } else if let value = value as? InitConfig {
+      super.writeByte(131)
       super.writeValue(value.toList())
     } else if let value = value as? PayConfig {
-      super.writeByte(129)
+      super.writeByte(132)
       super.writeValue(value.toList())
     } else {
       super.writeValue(value)
@@ -161,49 +293,52 @@ private class SberPayApiCodecWriter: FlutterStandardWriter {
   }
 }
 
-private class SberPayApiCodecReaderWriter: FlutterStandardReaderWriter {
+private class MessagesPigeonCodecReaderWriter: FlutterStandardReaderWriter {
   override func reader(with data: Data) -> FlutterStandardReader {
-    return SberPayApiCodecReader(data: data)
+    return MessagesPigeonCodecReader(data: data)
   }
 
   override func writer(with data: NSMutableData) -> FlutterStandardWriter {
-    return SberPayApiCodecWriter(data: data)
+    return MessagesPigeonCodecWriter(data: data)
   }
 }
 
-class SberPayApiCodec: FlutterStandardMessageCodec {
-  static let shared = SberPayApiCodec(readerWriter: SberPayApiCodecReaderWriter())
+class MessagesPigeonCodec: FlutterStandardMessageCodec, @unchecked Sendable {
+  static let shared = MessagesPigeonCodec(readerWriter: MessagesPigeonCodecReaderWriter())
 }
+
 
 /// Generated protocol from Pigeon that represents a handler of messages from Flutter.
 protocol SberPayApi {
-  func initSberPay(config: InitConfig) throws -> Bool
+  func initSberPay(config: InitConfig, completion: @escaping (Result<Bool, Error>) -> Void)
   func isReadyForSPaySdk() throws -> Bool
   func payWithBankInvoiceId(config: PayConfig, completion: @escaping (Result<SberPayApiPaymentStatus, Error>) -> Void)
 }
 
 /// Generated setup class from Pigeon to handle messages through the `binaryMessenger`.
 class SberPayApiSetup {
-  /// The codec used by SberPayApi.
-  static var codec: FlutterStandardMessageCodec { SberPayApiCodec.shared }
+  static var codec: FlutterStandardMessageCodec { MessagesPigeonCodec.shared }
   /// Sets up an instance of `SberPayApi` to handle messages through the `binaryMessenger`.
-  static func setUp(binaryMessenger: FlutterBinaryMessenger, api: SberPayApi?) {
-    let initSberPayChannel = FlutterBasicMessageChannel(name: "dev.flutter.pigeon.sber_pay_ios.SberPayApi.initSberPay", binaryMessenger: binaryMessenger, codec: codec)
+  static func setUp(binaryMessenger: FlutterBinaryMessenger, api: SberPayApi?, messageChannelSuffix: String = "") {
+    let channelSuffix = messageChannelSuffix.count > 0 ? ".\(messageChannelSuffix)" : ""
+    let initSberPayChannel = FlutterBasicMessageChannel(name: "dev.flutter.pigeon.sber_pay_ios.SberPayApi.initSberPay\(channelSuffix)", binaryMessenger: binaryMessenger, codec: codec)
     if let api = api {
       initSberPayChannel.setMessageHandler { message, reply in
         let args = message as! [Any?]
         let configArg = args[0] as! InitConfig
-        do {
-          let result = try api.initSberPay(config: configArg)
-          reply(wrapResult(result))
-        } catch {
-          reply(wrapError(error))
+        api.initSberPay(config: configArg) { result in
+          switch result {
+          case .success(let res):
+            reply(wrapResult(res))
+          case .failure(let error):
+            reply(wrapError(error))
+          }
         }
       }
     } else {
       initSberPayChannel.setMessageHandler(nil)
     }
-    let isReadyForSPaySdkChannel = FlutterBasicMessageChannel(name: "dev.flutter.pigeon.sber_pay_ios.SberPayApi.isReadyForSPaySdk", binaryMessenger: binaryMessenger, codec: codec)
+    let isReadyForSPaySdkChannel = FlutterBasicMessageChannel(name: "dev.flutter.pigeon.sber_pay_ios.SberPayApi.isReadyForSPaySdk\(channelSuffix)", binaryMessenger: binaryMessenger, codec: codec)
     if let api = api {
       isReadyForSPaySdkChannel.setMessageHandler { _, reply in
         do {
@@ -216,17 +351,17 @@ class SberPayApiSetup {
     } else {
       isReadyForSPaySdkChannel.setMessageHandler(nil)
     }
-    let payWithBankInvoiceIdChannel = FlutterBasicMessageChannel(name: "dev.flutter.pigeon.sber_pay_ios.SberPayApi.payWithBankInvoiceId", binaryMessenger: binaryMessenger, codec: codec)
+    let payWithBankInvoiceIdChannel = FlutterBasicMessageChannel(name: "dev.flutter.pigeon.sber_pay_ios.SberPayApi.payWithBankInvoiceId\(channelSuffix)", binaryMessenger: binaryMessenger, codec: codec)
     if let api = api {
       payWithBankInvoiceIdChannel.setMessageHandler { message, reply in
         let args = message as! [Any?]
         let configArg = args[0] as! PayConfig
         api.payWithBankInvoiceId(config: configArg) { result in
           switch result {
-            case .success(let res):
-              reply(wrapResult(res.rawValue))
-            case .failure(let error):
-              reply(wrapError(error))
+          case .success(let res):
+            reply(wrapResult(res))
+          case .failure(let error):
+            reply(wrapError(error))
           }
         }
       }
