@@ -1,6 +1,6 @@
 #if 0
 #elif defined(__arm64__) && __arm64__
-// Generated by Apple Swift version 6.0.3 effective-5.10 (swiftlang-6.0.3.1.10 clang-1600.0.30.1)
+// Generated by Apple Swift version 6.0 effective-5.10 (swiftlang-6.0.0.9.10 clang-1600.0.26.2)
 #ifndef SPAYSDK_SWIFT_H
 #define SPAYSDK_SWIFT_H
 #pragma clang diagnostic push
@@ -422,6 +422,7 @@ typedef SWIFT_ENUM_NAMED(NSInteger, SPayMethodObjc, "SPayMethodObjc", open) {
   SPayMethodObjcWithPaymentAccount = 4,
   SPayMethodObjcWithPartPay = 5,
   SPayMethodObjcWithBinding = 6,
+  SPayMethodObjcWithPhoneNumber = 7,
 };
 
 typedef SWIFT_ENUM(NSInteger, SPayState, open) {
@@ -502,7 +503,7 @@ SWIFT_CLASS_NAMED("SPaymentRequestObjc")
 #endif
 
 #elif defined(__x86_64__) && __x86_64__
-// Generated by Apple Swift version 6.0.3 effective-5.10 (swiftlang-6.0.3.1.10 clang-1600.0.30.1)
+// Generated by Apple Swift version 6.0 effective-5.10 (swiftlang-6.0.0.9.10 clang-1600.0.26.2)
 #ifndef SPAYSDK_SWIFT_H
 #define SPAYSDK_SWIFT_H
 #pragma clang diagnostic push
@@ -924,6 +925,7 @@ typedef SWIFT_ENUM_NAMED(NSInteger, SPayMethodObjc, "SPayMethodObjc", open) {
   SPayMethodObjcWithPaymentAccount = 4,
   SPayMethodObjcWithPartPay = 5,
   SPayMethodObjcWithBinding = 6,
+  SPayMethodObjcWithPhoneNumber = 7,
 };
 
 typedef SWIFT_ENUM(NSInteger, SPayState, open) {
